fixed button arret partie
